+++ models/CalendarEntry.php
@@ -18,6 +18,7 @@ use humhub\modules\space\models\Membership;
 use humhub\modules\space\models\Space;
 use humhub\modules\calendar\jobs\ForceParticipation;
 use humhub\widgets\Label;
+use Sabre\VObject\UUIDUtil;
 use Yii;
 use DateTime;
 use DateInterval;
@@ -46,6 +47,7 @@ use yii\db\ActiveQuery;
  * @property string $color
  * @property integer $external_source_id
  * @property string $external_id
+ * @property string $uid
  * @property integer $allow_decline
  * @property integer $allow_maybe
  * @property string $participant_info
@@ -214,6 +216,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      *
      * @param string $attribute attribute name
      * @param array $params parameters
+     * @throws \Exception
      */
     public function validateEndTime($attribute, $params)
     {
@@ -258,7 +261,16 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
             $this->all_day = 1;
         }
 
+        if($this->hasProperty('uid') && empty($this->uid)) {
+            $this->uid = static::createUUid();
+        }
+
         return parent::beforeSave($insert);
+    }
+
+    public static function createUUid($type = 'event')
+    {
+        return 'humhub-'.$type.'-' . UUIDUtil::getUUID();
     }
 
     public function beforeDelete()
@@ -285,6 +297,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
     /**
      * @param string $notificationClass
      * @throws \yii\base\InvalidConfigException
+     * @throws \Throwable
      */
     public function sendUpdateNotification($notificationClass = EventUpdated::class)
     {
@@ -380,6 +393,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      *
      * @param User $user
      * @return CalendarEntryParticipant
+     * @throws \Throwable
      */
     public function findParticipant(User $user = null)
     {
@@ -444,6 +458,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      *
      * @param User $user
      * @return boolean
+     * @throws \Throwable
      */
     public function canRespond(User $user = null)
     {
@@ -486,6 +501,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      *
      * @param User $user
      * @return boolean
+     * @throws \Throwable
      */
     public function hasResponded(User $user = null)
     {
@@ -631,6 +647,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      * @param int $limit
      * @return CalendarEntry[]
      * @throws Exception
+     * @throws \Throwable
      * @see CalendarEntryQuery
      */
     public static function getEntriesByRange(DateTime $start, DateTime $end, $includes = [], $filters = [], $limit = 50)
@@ -655,6 +672,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
      * @param int $daysInFuture
      * @param int $limit
      * @return CalendarEntry[]
+     * @throws \Throwable
      */
     public static function getUpcomingEntries(ContentContainerActiveRecord $contentContainer = null, $daysInFuture = 7, $limit = 5)
     {
@@ -723,5 +741,46 @@ class CalendarEntry extends ContentActiveRecord implements Searchable, CalendarI
             }
         }
 
+    }
+
+    /**
+     * @return string
+     */
+    public function getUid()
+    {
+        return $this->uid;
+    }
+
+    public function setUid($uid)
+    {
+        $this->uid = $uid;
+    }
+
+    /**
+     * @return boolean
+     */
+    public function isExportable()
+    {
+        return true;
+    }
+
+    public function getRrule()
+    {
+        return null;
+    }
+
+    public function getExdate()
+    {
+        return null;
+    }
+
+    public function getLocation()
+    {
+        return null;
+    }
+
+    public function getDescription()
+    {
+        return $this->description;
     }
 }