+++ widgets/views/wallEntry.php
@@ -1,8 +1,7 @@
 <?php
 
 use humhub\modules\calendar\widgets\EntryParticipants;
-use humhub\widgets\Button;
-use humhub\widgets\MarkdownView;
+use humhub\modules\content\widgets\richtext\RichText;
 use yii\helpers\Html;
 use humhub\modules\calendar\models\CalendarEntryParticipant;
 use humhub\modules\calendar\models\CalendarEntry;
@@ -15,7 +14,7 @@ $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable(
 ?>
 
 <div class="media event" style="" data-action-component="calendar.CalendarEntry" data-calendar-entry="<?= $calendarEntry->id ?>">
-    <div class="y" style="padding-left:10px; border-left: 3px solid <?= $color ?>">
+    <div style="padding-left:10px; border-left: 3px solid <?= $color ?>">
         <div class="media-body clearfix">
             <a href="<?= $calendarEntry->getUrl(); ?>" class="pull-left" style="margin-right: 10px">
                 <i class="fa fa-calendar colorDefault" style="font-size: 35px;"></i>
@@ -31,7 +30,7 @@ $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable(
         </div>
         <?php if (!empty($calendarEntry->description)) : ?>
             <div <?= ($collapse) ? 'data-ui-show-more' : '' ?> data-read-more-text="<?= Yii::t('CalendarModule.views_entry_view', "Read full description...") ?>" style="overflow:hidden">
-                <?= MarkdownView::widget(['markdown' => $calendarEntry->description]); ?>
+                <?= RichText::output($calendarEntry->description); ?>
             </div>
         <?php endif; ?>
     </div>
@@ -43,7 +42,7 @@ $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable(
                 <div class="col-md-12">
                     <div <?= ($collapse) ? 'data-ui-show-more' : '' ?> data-read-more-text="<?= Yii::t('CalendarModule.views_entry_view', "Read full participation info...") ?>">
                         <strong><i class="fa fa-info-circle"></i> <?= Yii::t('CalendarModule.views_entry_view', 'Participant information:') ?></strong>
-                        <?= MarkdownView::widget(['markdown' => $calendarEntry->participant_info]); ?>
+                        <?= RichText::output( $calendarEntry->participant_info); ?>
                     </div>
                 </div>
             </div>
@@ -62,7 +61,6 @@ $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable(
     <?php if ($calendarEntry->canRespond()): ?>
        <div class="row" style="padding-top:10px">
             <div class="col-md-12">
-
               <?= EntryParticipants::participateButton($calendarEntry, CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED, Yii::t('CalendarModule.views_entry_view', "Attend")); ?>
               <?= EntryParticipants::participateButton($calendarEntry, CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE,    Yii::t('CalendarModule.views_entry_view', "Maybe")); ?>
               <?= EntryParticipants::participateButton($calendarEntry, CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED, Yii::t('CalendarModule.views_entry_view', "Decline")); ?>